Combined ESS "Reason" and "Notification" messages into a single Marquee
ESS "reason" codes were originally reported as a string of numbers (e.g,: #1 #2). You had to go look up the numbers to know what was being reported. In addition, "notifications" were displayed on a sparate line as a scrolling "Marquee". This pushed the clock off the bottom of the tile.

I combined the ESS reason and notifications on a single Marquee line. I also replaced the numbers with a text description of the reason.

I also commented out Reason #2 Battery Life because I found it distracting. You can easily reactivate reporting this in SystemReasonMessage.qml

For these changes, you must replace OverviewMobile.qml and ADD SystemReasonMessage.qml. The latter was cloned from SystemReason.qml and modified to display text rather than the numbers.